Hopefully fix issue with Runtime error
This error seems to happen when the arango cursor times out, once
upon a time the texts were actually stored in arangoDB so keeping
the cursor open made sense to reduce memory use.

Now only the paths are stored in arango, so there's no reason to not
fetch all the results immediately and close the cursor.