Remove dependency on pali_concord.json
Instead, use `all_pali_concordance.json`.

Before, the PaliPageNumbinator loaded the contents of `pali_concord.json` into a dictionary that maps a triple—consisting of the Mahasangiti volume code, paragraph number, and the manuscript edition code—to a couple consisting of the book and page number in the PTS manuscripts for the corresponding volume.

After this change, PaliPageNumbinator loads the contents of `all_pali_concordance.json` into a dictionary as before, but the key is a couple, omitting the manuscript edition code which the new loading algorithm will ensure is always a PTS edition. The loading algorithm also handles which PTS edition to use, always preferring the latest one.