Replace print_stage() with StagePrinter.
In the short term this class gives us a way of monitoring the arangoload.run() function in our unit tests. It will likely evolve into something more capable, such as giving us timing information and so forth. It might also just get refactored out of existence as it's a pretty coarse thing.